+++ SensateIoT.Platform.Network.DataAccess/Abstract/ITriggerRepository.cs
@@ -1,10 +1,15 @@
-﻿using System.Collections.Generic;
+﻿/*
+ * Trigger service repository.
+ *
+ * @author Michel Megens
+ * @email  michel@michelmegens.net
+ */
+
+using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
 using MongoDB.Bson;
-
-using SensateIoT.Platform.Network.Data.Models;
 
 using TriggerAction = SensateIoT.Platform.Network.Data.DTO.TriggerAction;
 
@@ -12,7 +17,7 @@ namespace SensateIoT.Platform.Network.DataAccess.Abstract
 {
 	public interface ITriggerRepository
 	{
-		Task<IEnumerable<TriggerAction>> GetTriggerServiceActions(IEnumerable<ObjectId> sensorIds, CancellationToken ct = default);
-		Task StoreTriggerInvocation(TriggerInvocation invocation, CancellationToken ct = default);
+		Task<IEnumerable<TriggerAction>> GetTriggerServiceActions(CancellationToken ct = default);
+		Task<IEnumerable<TriggerAction>> GetTriggerServiceActionsBySensorId(ObjectId id, CancellationToken ct = default);
 	}
 }