+++ SensateIoT.Platform.Network.DataAccess/Repositories/TriggerRepository.cs
@@ -5,21 +5,19 @@
  * @email  michel@michelmegens.net
  */
 
-using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Threading;
 using System.Threading.Tasks;
 
 using MongoDB.Bson;
-
 using Npgsql;
 using NpgsqlTypes;
 
 using SensateIoT.Platform.Network.Data.DTO;
-using SensateIoT.Platform.Network.Data.Models;
 using SensateIoT.Platform.Network.DataAccess.Abstract;
-using SensateIoT.Platform.Network.DataAccess.Extensions;
+
+using TriggerAction = SensateIoT.Platform.Network.Data.DTO.TriggerAction;
 
 namespace SensateIoT.Platform.Network.DataAccess.Repositories
 {
@@ -28,16 +26,16 @@ namespace SensateIoT.Platform.Network.DataAccess.Repositories
 		private readonly INetworkingDbContext m_ctx;
 
 		private const string TriggerService_GetTriggersBySensorID = "triggerservice_gettriggersbysensorid";
-		private const string TriggerService_CreateInvocation = "triggerservice_createinvocation";
+		private const string TriggerService_GetTriggers = "triggerservice_gettriggers";
 
 		public TriggerRepository(INetworkingDbContext ctx)
 		{
 			this.m_ctx = ctx;
 		}
 
-		public async Task<IEnumerable<Data.DTO.TriggerAction>> GetTriggerServiceActions(IEnumerable<ObjectId> sensorIds, CancellationToken ct)
+		public async Task<IEnumerable<TriggerAction>> GetTriggerServiceActions(CancellationToken ct)
 		{
-			var result = new List<Data.DTO.TriggerAction>();
+			var result = new List<TriggerAction>();
 
 			await using var cmd = this.m_ctx.Connection.CreateCommand();
 
@@ -45,14 +43,12 @@ namespace SensateIoT.Platform.Network.DataAccess.Repositories
 				await cmd.Connection.OpenAsync(ct).ConfigureAwait(false);
 			}
 
-			var idArray = string.Join(",", sensorIds);
 			cmd.CommandType = CommandType.StoredProcedure;
-			cmd.CommandText = TriggerService_GetTriggersBySensorID;
-			cmd.Parameters.Add(new NpgsqlParameter("idlist", DbType.String) { Value = idArray });
+			cmd.CommandText = TriggerService_GetTriggers;
 
 			await using var reader = await cmd.ExecuteReaderAsync(ct).ConfigureAwait(false);
 			while(await reader.ReadAsync(ct)) {
-				var record = new Data.DTO.TriggerAction {
+				var record = new TriggerAction {
 					TriggerID = reader.GetInt64(0),
 					ActionID = reader.GetInt64(1),
 					SensorID = ObjectId.Parse(reader.GetString(2)),
@@ -61,8 +57,7 @@ namespace SensateIoT.Platform.Network.DataAccess.Repositories
 					Type = (TriggerType)reader.GetFieldValue<int>(7),
 					Channel = (TriggerChannel)reader.GetFieldValue<int>(8),
 					Target = !await reader.IsDBNullAsync(9, ct) ? reader.GetString(9) : null,
-					Message = !await reader.IsDBNullAsync(10, ct) ? reader.GetString(10) : null,
-					LastInvocation = !await reader.IsDBNullAsync(11, ct) ? reader.GetDateTime(11) : DateTime.MinValue
+					Message = !await reader.IsDBNullAsync(10, ct) ? reader.GetString(10) : null
 				};
 
 				if(await reader.IsDBNullAsync(4, ct)) {
@@ -83,18 +78,54 @@ namespace SensateIoT.Platform.Network.DataAccess.Repositories
 			return result;
 		}
 
-		public async Task StoreTriggerInvocation(TriggerInvocation invocation, CancellationToken ct)
+		public async Task<IEnumerable<TriggerAction>> GetTriggerServiceActionsBySensorId(ObjectId id, CancellationToken ct = default)
 		{
-			using var builder = StoredProcedureBuilder.Create(this.m_ctx.Connection);
+			var result = new List<TriggerAction>();
 
-			builder.WithParameter("triggerid", invocation.TriggerID, NpgsqlDbType.Bigint);
-			builder.WithParameter("actionid", invocation.ActionID, NpgsqlDbType.Bigint);
-			builder.WithParameter("timestmp", invocation.Timestamp, NpgsqlDbType.Timestamp);
+			await using var cmd = this.m_ctx.Connection.CreateCommand();
 
-			builder.WithFunction(TriggerService_CreateInvocation);
+			if(cmd.Connection != null && cmd.Connection.State != ConnectionState.Open) {
+				await cmd.Connection.OpenAsync(ct).ConfigureAwait(false);
+			}
 
-			var reader = await builder.ExecuteAsync(ct).ConfigureAwait(false);
-			await reader.DisposeAsync().ConfigureAwait(false);
+			cmd.CommandType = CommandType.StoredProcedure;
+			cmd.CommandText = TriggerService_GetTriggersBySensorID;
+			var param = new NpgsqlParameter("id", NpgsqlDbType.Varchar) {
+				Value = id.ToString()
+			};
+			cmd.Parameters.Add(param);
+
+			await using var reader = await cmd.ExecuteReaderAsync(ct).ConfigureAwait(false);
+			while(await reader.ReadAsync(ct)) {
+				var record = new TriggerAction {
+					TriggerID = reader.GetInt64(0),
+					ActionID = reader.GetInt64(1),
+					SensorID = ObjectId.Parse(reader.GetString(2)),
+					KeyValue = reader.GetString(3),
+					FormalLanguage = !await reader.IsDBNullAsync(6, ct) ? reader.GetString(6) : null,
+					Type = (TriggerType)reader.GetFieldValue<int>(7),
+					Channel = (TriggerChannel)reader.GetFieldValue<int>(8),
+					Target = !await reader.IsDBNullAsync(9, ct) ? reader.GetString(9) : null,
+					Message = !await reader.IsDBNullAsync(10, ct) ? reader.GetString(10) : null
+				};
+
+				if(await reader.IsDBNullAsync(4, ct)) {
+					record.LowerEdge = null;
+				} else {
+					record.LowerEdge = reader.GetDecimal(4);
+				}
+
+				if(await reader.IsDBNullAsync(5, ct)) {
+					record.UpperEdge = null;
+				} else {
+					record.UpperEdge = reader.GetDecimal(5);
+				}
+
+				result.Add(record);
+			}
+
+			return result;
+
 		}
 	}
 }