+++ SensateIoT.Platform.Network.Router/Application/Startup.cs
@@ -14,21 +14,11 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
 using Prometheus;
+using JetBrains.Annotations;
 
-using SensateIoT.Platform.Network.Common.Caching.Abstract;
-using SensateIoT.Platform.Network.Common.Caching.Routing;
-using SensateIoT.Platform.Network.Common.Collections.Abstract;
-using SensateIoT.Platform.Network.Common.Collections.Local;
-using SensateIoT.Platform.Network.Common.Collections.Remote;
 using SensateIoT.Platform.Network.Common.Init;
-using SensateIoT.Platform.Network.Common.Services.Data;
-using SensateIoT.Platform.Network.Common.Services.Metrics;
-using SensateIoT.Platform.Network.Common.Services.Processing;
-using SensateIoT.Platform.Network.Common.Settings;
-using SensateIoT.Platform.Network.Data.Abstract;
-using SensateIoT.Platform.Network.DataAccess.Abstract;
-using SensateIoT.Platform.Network.DataAccess.Repositories;
 using SensateIoT.Platform.Network.Router.Config;
+using SensateIoT.Platform.Network.Router.Init;
 using SensateIoT.Platform.Network.Router.MQTT;
 using SensateIoT.Platform.Network.Router.Services;
 
@@ -43,94 +33,29 @@ namespace SensateIoT.Platform.Network.Router.Application
 			this.Configuration = configuration;
 		}
 
+		[UsedImplicitly]
 		public void ConfigureServices(IServiceCollection services)
 		{
 			var db = new DatabaseConfig();
-			var mqtt = new MqttConfig();
 
 			this.Configuration.GetSection("Database").Bind(db);
-			this.Configuration.GetSection("Mqtt").Bind(mqtt);
-
-			var id = this.Configuration.GetValue<string>("ApplicationId");
-			var reload = this.Configuration.GetValue<int>("Cache:DataReloadInterval");
-			var privatemqtt = mqtt.InternalBroker;
-			var publicmqtt = mqtt.PublicBroker;
 
 			services.AddDocumentStore(db.MongoDB.ConnectionString, db.MongoDB.DatabaseName, db.MongoDB.MaxConnections);
 			services.AddConnectionStrings(db.Networking.ConnectionString, db.SensateIoT.ConnectionString);
 			services.AddAuthorizationContext();
 			services.AddNetworkingContext();
-
-			services.Configure<DataReloadSettings>(opts => {
-				opts.StartDelay = TimeSpan.FromSeconds(1);
-				opts.EnableReload = this.Configuration.GetValue<bool>("Cache:EnableReload");
-
-				/* Default to 30 minutes for live data handler reload */
-				opts.DataReloadInterval = reload == 0 ? TimeSpan.FromMinutes(30) : TimeSpan.FromSeconds(reload);
-				opts.LiveDataReloadInterval = TimeSpan.FromSeconds(this.Configuration.GetValue<int>("Cache:LiveDataReloadInterval"));
-				opts.TimeoutScanInterval = TimeSpan.FromSeconds(this.Configuration.GetValue<int>("Cache:TimeoutScanInterval"));
-			});
-
-			services.Configure<QueueSettings>(s => {
-				s.LiveDataQueueTemplate = this.Configuration.GetValue<string>("Routing:LiveDataTopic");
-				s.TriggerQueueTemplate = this.Configuration.GetValue<string>("Routing:TriggerTopic");
-				s.MessageStorageQueueTopic = this.Configuration.GetValue<string>("Routing:MessageStorageQueueTopic");
-				s.MeasurementStorageQueueTopic = this.Configuration.GetValue<string>("Routing:MeasurementStorageQueueTopic");
-				s.NetworkEventQueueTopic = this.Configuration.GetValue<string>("Routing:NetworkEventQueueTopic");
-			});
-
-			services.Configure<RoutingPublishSettings>(s => {
-				s.InternalInterval = TimeSpan.FromMilliseconds(this.Configuration.GetValue<int>("Routing:InternalPublishInterval"));
-				s.PublicInterval = TimeSpan.FromMilliseconds(this.Configuration.GetValue<int>("Routing:PublicPublishInterval"));
-				s.ActuatorTopicFormat = this.Configuration.GetValue<string>("Routing:ActuatorTopicFormat");
-			});
-
-			services.Configure<MetricsOptions>(this.Configuration.GetSection("HttpServer:Metrics"));
-
-			services.AddInternalMqttService(options => {
-				options.Ssl = privatemqtt.Ssl;
-				options.Host = privatemqtt.Host;
-				options.Port = privatemqtt.Port;
-				options.Username = privatemqtt.Username;
-				options.Password = privatemqtt.Password;
-				options.Id = $"router-{id}";
-			});
-
-			services.AddMqttService(options => {
-				options.Ssl = publicmqtt.Ssl;
-				options.Host = publicmqtt.Host;
-				options.Port = publicmqtt.Port;
-				options.Username = publicmqtt.Username;
-				options.Password = publicmqtt.Password;
-				options.Id = $"router-{id}-{Guid.NewGuid():N}";
-			});
-
-			services.AddScoped<IRoutingRepository, RoutingRepository>();
-			services.AddScoped<ILiveDataHandlerRepository, LiveDataHandlerRepository>();
+			services.AddRoutingServices(this.Configuration);
+			services.AddBackgroundServices(this.Configuration);
+			services.AddMqttBrokers(this.Configuration);
 
 			services.AddSingleton<CommandCounter>();
-			services.AddSingleton<IQueue<IPlatformMessage>, Deque<IPlatformMessage>>();
-			services.AddSingleton<IMessageQueue, MessageQueue>();
-			services.AddSingleton<IRemoteNetworkEventQueue, RemoteNetworkEventQueue>();
-			services.AddSingleton<IInternalRemoteQueue, InternalMqttQueue>();
-			services.AddSingleton<IPublicRemoteQueue, PublicMqttQueue>();
-			services.AddSingleton<IAuthorizationService, AuthorizationService>();
-			services.AddSingleton<IRemoteStorageQueue, RemoteStorageQueue>();
-			services.AddSingleton<IRoutingCache, RoutingCache>();
-
-			services.AddSingleton<IHostedService, DataReloadService>();
-			services.AddSingleton<IHostedService, CacheTimeoutScanService>();
-			services.AddSingleton<IHostedService, LiveDataReloadService>();
-			services.AddSingleton<IHostedService, RoutingPublishService>();
-			services.AddSingleton<IHostedService, ActuatorPublishService>();
-			services.AddSingleton<IHostedService, RoutingService>();
-			services.AddSingleton<IHostedService, MetricsService>();
 
 			services.AddGrpc();
 			services.AddGrpcReflection();
 			services.AddMqttHandlers();
 		}
 
+		[UsedImplicitly]
 		public void Configure(IApplicationBuilder app, IWebHostEnvironment env, IServiceProvider provider)
 		{
 			var mqtt = new MqttConfig();