+++ SensateIoT.Platform.Network.TriggerService/Application/Startup.cs
@@ -21,6 +21,8 @@ using SensateIoT.Platform.Network.Common.Services.Metrics;
 using SensateIoT.Platform.Network.Common.Settings;
 using SensateIoT.Platform.Network.DataAccess.Abstract;
 using SensateIoT.Platform.Network.DataAccess.Repositories;
+using SensateIoT.Platform.Network.TriggerService.Abstract;
+using SensateIoT.Platform.Network.TriggerService.Caching;
 using SensateIoT.Platform.Network.TriggerService.Clients;
 using SensateIoT.Platform.Network.TriggerService.Config;
 using SensateIoT.Platform.Network.TriggerService.MQTT;
@@ -61,6 +63,7 @@ namespace SensateIoT.Platform.Network.TriggerService.Application
 			services.AddDocumentStore(db.MongoDB.ConnectionString, db.MongoDB.DatabaseName, db.MongoDB.MaxConnections);
 			services.Configure<TimeoutConfig>(this.Configuration.GetSection("Timeouts"));
 
+			services.Configure<MqttConfig>(this.Configuration.GetSection("Mqtt"));
 			services.Configure<MetricsOptions>(this.Configuration.GetSection("HttpServer:Metrics"));
 			services.AddHostedService<MetricsService>();
 
@@ -112,7 +115,7 @@ namespace SensateIoT.Platform.Network.TriggerService.Application
 
 				services.AddScoped<ITextSendService, TwillioTextSendService>();
 			} else {
-				Console.WriteLine("Text message provider not configured!");
+				throw new InvalidOperationException("Text provider not configured!");
 			}
 
 			services.AddScoped<IControlMessageRepository, ControlMessageRepository>();
@@ -120,10 +123,14 @@ namespace SensateIoT.Platform.Network.TriggerService.Application
 			services.AddScoped<IRoutingRepository, RoutingRepository>();
 			services.AddScoped<IControlMessageRepository, ControlMessageRepository>();
 			services.AddScoped<ITriggerActionExecutionService, TriggerActionExecutionService>();
+
 			services.AddSingleton<IDataPointMatchingService, DataPointMatchingService>();
 			services.AddSingleton<IRegexMatchingService, RegexMatchingService>();
 			services.AddSingleton<IRouterClient, RouterClient>();
 			services.AddSingleton<IEmailSender, SmtpMailer>();
+			services.AddSingleton<ITriggerActionCache, TriggerActionCache>();
+
+			services.AddHostedService<TriggerActionReloadService>();
 
 			services.AddHttpClient();
 			services.AddMqttHandlers();
@@ -138,6 +145,7 @@ namespace SensateIoT.Platform.Network.TriggerService.Application
 
 			provider.MapInternalMqttTopic<MqttBulkNumberTriggerHandler>(@private.BulkMeasurementTopic);
 			provider.MapInternalMqttTopic<MqttRegexTriggerHandler>(@private.BulkMessageTopic);
+			provider.MapInternalMqttTopic<CommandConsumer>(@private.CommandTopic);
 		}
 	}
 }